+++ src/Security/Security/ChangeLog.md
@@ -18,7 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Add new cmdlets: 'Get-AzSecurityTopologies'
+* Add new cmdlets: 'Get-AzSecurityTopology'
 
 ## Version 0.7.9
 * Added new cmdlets: